fix add Bigdecimal

+++ src/main/java/by/intereson/ebookservice/entities/Book.java
@@ -8,6 +8,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import static jakarta.persistence.EnumType.STRING;
@@ -26,27 +27,38 @@ public class Book {
     @GeneratedValue(strategy = SEQUENCE, generator = SEQ_NAME)
     @SequenceGenerator(name = SEQ_NAME, sequenceName = SEQ_NAME, allocationSize = 1, initialValue = 100)
     private Long id;
+
     @Column(name = "BOOK_NAME", nullable = false)
     private String bookName;
+
     @Column(name = "AUTHOR", nullable = false)
     private String author;
+
     @Column(name = "PUBLISHING_YEAR")
     private Integer publishingYear;
+
     @Column(name = "PUBLISHING_HOUSE")
     private String publishingHouse;
+
     @Column(name = "DESCRIPTION")
     private String description;
-    @Column(name = "PRICE",nullable = false)
-    private Double price;
+
+    @Column(name = "PRICE", nullable = false)
+    private BigDecimal price;
+
     @Column(name = "CREATE_TIME")
     @CreationTimestamp
     private String dateTime;
+
     @Enumerated(STRING)
     private Genre genre;
-    @Column(name = "QUANTITY",nullable = false)
+
+    @Column(name = "QUANTITY", nullable = false)
     private Integer quantity;
+
     @Column(name = "RESERVE_QUANTITY")
     private Integer reserveQuantity;
+
     @ManyToMany(mappedBy = "likedBooks")
     private List<User> users;
